feat: Add login redirect for unauthorized users
This commit adds a login redirect for unauthorized users in the DoctorPatientsPage and PatientLayout components. If the user is not logged in, they will be redirected to the login page. Additionally, if the user's role is "patient" in the DoctorPatientsPage component or "doctor" in the PatientLayout component, they will be redirected to the appropriate protected page.

Note: This commit message follows the established convention of using the "feat" type for new features or enhancements.