refactor: Update DoctorOnboarding component to add doctor to hospital's list of doctors
This commit refactors the DoctorOnboarding component to add the newly onboarded doctor to the hospital's list of doctors. The changes include updating the hospital data with the doctor's information and merging it with the existing hospital data. This ensures that the doctor is associated with the correct hospital and can be displayed in the hospital's list of doctors.

Note: This commit message follows the established convention of using the "refactor" type for code refactoring and improvement.