feat: Add tooltip to display hospital number in profile page
This commit adds a tooltip to the profile page that displays the hospital number. The tooltip is triggered when the user hovers over the hospital number badge. This enhancement provides a convenient way for users to copy the hospital number to their clipboard.

Note: This commit message follows the established convention of using the "feat" type for new features.